fixes requested by Marcin
