+++ Wire-iOS/Sources/UserInterface/Settings/CellDescriptors/SettingsCellDescriptorFactory+Account.swift
@@ -282,6 +282,7 @@ extension SettingsCellDescriptorFactory {
             title: "self.settings.account_picture_group.color".localized,
             isDestructive: false,
             presentationStyle: .navigation,
+            identifier: nil,
             presentationAction: AccentColorPickerController.init,
             previewGenerator: { _ in .color(ZMUser.selfUser().accentColor) }
         )
