+++ Wire-iOS/Sources/UserInterface/Settings/CellDescriptors/SettingsExternalScreenColorCellDescriptor.swift
@@ -1,38 +1,26 @@
-// 
+//
 // Wire
-// Copyright (C) 2016 Wire Swiss GmbH
-// 
+// Copyright (C) 2022 Wire Swiss GmbH
+//
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
 // the Free Software Foundation, either version 3 of the License, or
 // (at your option) any later version.
-// 
+//
 // This program is distributed in the hope that it will be useful,
 // but WITHOUT ANY WARRANTY; without even the implied warranty of
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 // GNU General Public License for more details.
-// 
+//
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see http://www.gnu.org/licenses/.
-// 
+//
 
-import Foundation
 import UIKit
 import WireCommonComponents
 
-enum PresentationStyle: Int {
-    case modal
-    case navigation
-}
-
-enum AccessoryViewMode: Int {
-    case `default`
-    case alwaysShow
-    case alwaysHide
-}
-
-class SettingsExternalScreenCellDescriptor: SettingsExternalScreenCellDescriptorType, SettingsControllerGeneratorType {
-    static let cellType = SettingsTableCell.self
+class SettingsExternalScreenColorCellDescriptor: SettingsExternalScreenCellDescriptorType, SettingsControllerGeneratorType {
+    static let cellType: SettingsTableCell.Type = SettingsTableColorCell.self
     var visible: Bool = true
     let title: String
     let destructive: Bool
@@ -48,37 +36,6 @@ class SettingsExternalScreenCellDescriptor: SettingsExternalScreenCellDescriptor
     let previewGenerator: PreviewGeneratorType?
 
     let presentationAction: () -> (UIViewController?)
-
-    convenience init(title: String, presentationAction: @escaping () -> (UIViewController?)) {
-        self.init(
-            title: title,
-            isDestructive: false,
-            presentationStyle: .navigation,
-            identifier: nil,
-            presentationAction: presentationAction,
-            previewGenerator: nil,
-            icon: .none
-        )
-    }
-
-    convenience init(title: String,
-                     isDestructive: Bool,
-                     presentationStyle: PresentationStyle,
-                     presentationAction: @escaping () -> (UIViewController?),
-                     previewGenerator: PreviewGeneratorType? = .none,
-                     icon: StyleKitIcon? = nil,
-                     accessoryViewMode: AccessoryViewMode = .default) {
-        self.init(
-            title: title,
-            isDestructive: isDestructive,
-            presentationStyle: presentationStyle,
-            identifier: nil,
-            presentationAction: presentationAction,
-            previewGenerator: previewGenerator,
-            icon: icon,
-            accessoryViewMode: accessoryViewMode
-        )
-    }
 
     init(title: String,
          isDestructive: Bool,
