+++ Wire-iOS/Sources/UserInterface/Settings/SettingsTableColorCell.swift
@@ -117,15 +117,3 @@ class SettingsTableColorCell: SettingsTableCell {
         }
     }
 }
-
-extension UIColor {
-    func hexValue() -> String {
-        let components = self.cgColor.components
-        let r: CGFloat = components?[0] ?? 0.0
-        let g: CGFloat = components?[1] ?? 0.0
-        let b: CGFloat = components?[2] ?? 0.0
-
-        let hexString = String.init(format: "#%02lX%02lX%02lX", lroundf(Float(r * 255)), lroundf(Float(g * 255)), lroundf(Float(b * 255)))
-        return hexString
-     }
-}
